# swagger.yml
basePath: /
consumes:
    - application/json
definitions:
...
    AuthParams:
        description: AuthParams - struct for auth params
        properties:
            id:
                type: string
                x-go-name: ID
            macaddress:
                type: string
                x-go-name: MacAddress
            password:
                type: string
                x-go-name: Password
        type: object
        x-go-package: github.com/gravitl/netmaker/models
...
    SuccessResponse:
        properties:
            Code:
                format: int64
                type: integer
            Message:
                type: string
            Response: {}
        title: SuccessResponse is struct for sending error message with code.
        type: object
        x-go-package: github.com/gravitl/netmaker/models
...
    User:
        description: User struct - struct for Users
        properties:
            isadmin:
                type: boolean
                x-go-name: IsAdmin
            issuperadmin:
                type: boolean
                x-go-name: IsSuperAdmin
            password:
                type: string
                x-go-name: Password
            remote_gw_ids:
                additionalProperties:
                    type: object
                type: object
                x-go-name: RemoteGwIDs
            username:
                type: string
                x-go-name: UserName
        type: object
        x-go-package: github.com/gravitl/netmaker/models
    UserAuthParams:
        description: UserAuthParams - user auth params struct
        properties:
            password:
                type: string
                x-go-name: Password
            username:
                type: string
                x-go-name: UserName
        type: object
        x-go-package: github.com/gravitl/netmaker/models
host: api.demo.netmaker.io
info:
    description: |-
        # API Usage

        Most actions that can be performed via API can be performed via UI. We recommend managing your networks using the official netmaker-ui project. However, Netmaker can also be run without the UI, and all functions can be achieved via API calls. If your use case requires using Netmaker without the UI or you need to do some troubleshooting/advanced configuration, using the API directly may help.

        # Authentication

        API calls must be authenticated via a header of the format -H “Authorization: Bearer <YOUR_SECRET_KEY>” There are two methods to obtain YOUR_SECRET_KEY: 1. Using the masterkey. By default, this value is “secret key,” but you should change this on your instance and keep it secure. This value can be set via env var at startup or in a config file (config/environments/< env >.yaml). See the [Netmaker](https://docs.netmaker.org/index.html) documentation for more details. 2. Using a JWT received for a node. This can be retrieved by calling the /api/nodes/<network>/authenticate endpoint, as documented below.
    title: Netmaker
    version: 0.21.1
paths:
...
    /api/nodes/adm/{network}/authenticate:
        post:
            operationId: authenticate
            parameters:
                - description: network
                  in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
                - description: AuthParams
                  in: body
                  name: auth_params
                  schema:
                    $ref: '#/definitions/AuthParams'
                  x-go-name: AuthParams
            responses:
                "200":
                    $ref: '#/responses/successResponse'
            schemes:
                - https
            summary: Authenticate to make further API calls related to a network.
            tags:
                - authenticate
...
    /api/users/{username}:
        delete:
            operationId: deleteUser
            parameters:
                - description: Username
                  in: path
                  name: username
                  required: true
                  type: string
                  x-go-name: Username
            responses:
                "200":
                    $ref: '#/responses/userBodyResponse'
            schemes:
                - https
            summary: Delete a user.
            tags:
                - user
        get:
            operationId: getUser
            parameters:
                - description: Username
                  in: path
                  name: username
                  required: true
                  type: string
                  x-go-name: Username
            responses:
                "200":
                    $ref: '#/responses/userBodyResponse'
            schemes:
                - https
            summary: Get an individual user.
            tags:
                - user
        post:
            operationId: createUser
            parameters:
                - description: User
                  in: body
                  name: user
                  schema:
                    $ref: '#/definitions/User'
                  x-go-name: User
                - description: Username
                  in: path
                  name: username
                  required: true
                  type: string
                  x-go-name: Username
            responses:
                "200":
                    $ref: '#/responses/userBodyResponse'
            schemes:
                - https
            summary: Create a user.
            tags:
                - user
        put:
            operationId: updateUser
            parameters:
                - description: User
                  in: body
                  name: user
                  schema:
                    $ref: '#/definitions/User'
                  x-go-name: User
                - description: Username
                  in: path
                  name: username
                  required: true
                  type: string
                  x-go-name: Username
            responses:
                "200":
                    $ref: '#/responses/userBodyResponse'
            schemes:
                - https
            summary: Update a user.
            tags:
                - user
...
    /api/users/adm/authenticate:
        post:
            operationId: authenticateUser
            parameters:
                - description: User Auth Params
                  in: body
                  name: user_auth_params
                  schema:
                    $ref: '#/definitions/UserAuthParams'
                  x-go-name: UserAuthParams
            responses:
                "200":
                    $ref: '#/responses/successResponse'
            schemes:
                - https
            summary: User authenticates using its password and retrieves a JWT for authorization.
            tags:
                - authenticate
    /api/users/adm/createsuperadmin:
        post:
            operationId: createAdmin
            parameters:
                - description: User
                  in: body
                  name: user
                  schema:
                    $ref: '#/definitions/User'
                  x-go-name: User
            responses:
                "200":
                    $ref: '#/responses/userBodyResponse'
            schemes:
                - https
            summary: Make a user an admin.
            tags:
                - user
...
produces:
    - application/json
responses:
...
    successResponse:
        description: ""
        schema:
            $ref: '#/definitions/SuccessResponse'
    userBodyResponse:
        description: ""
        schema:
            $ref: '#/definitions/User'
schemes:
    - https
swagger: "2.0"